[usb setting logic] Do not check for mode_to_set_twice Fixes: CID#1230272
In case we compile without devicelock support the old code already checked
for a mode to set. However this meant the secondary part of the loop got never
reached. Now we check in one place. No more dead code. Also we cleaned-up a
minor mem leak at the same time, as the mode_to_set did not get cleaned-up in
case it matched the current mode.

Signed-off-by: Philippe De Swert <philippe.deswert@jollamobile.com>