[debouncing] Improve debouncing to avoid missing signals. Fixes: JB#25557
In some border line cases there can be issues with signals not being sent.
This is the case when a charger gets identified as cable and later on correctly
as a dedicated charger.
We make sure we actually compare the set state instead of just relying on
connected state only.

Signed-off-by: Philippe De Swert <philippe.deswert@jollamobile.com>