[cleanup] Remove utils/udev-search.c
This file was never built and duplicated code that is in
src/usb_moded-udev.c. If we wanted to have this utility,
it should re-use the code in usb-moded-udev.c (make it modular
and callable from the udev-search.c main() without duplicating
code) and be built unconditionally (and then put into some -tools
or -rnd-tools package).