[network] Improve dealing with missing interfaces.
When the interface is missing or not yet up we try to
gracefully give it a second try. As sometimes something
like a functionfs using gadget could take a while to get
initialized. Also do not fail on a failure to set up the
dhcp server. Falling back on manual setup is better than
not being able to do so.

Signed-off-by: Philippe De Swert <philippe.deswert@jollamobile.com>